refactor(*): 패키지 위치를 정리
