+++ src/main/java/lotto/lottoapp/model/Lotto.java
@@ -2,9 +2,9 @@ package lotto.lottoapp.model;
 
 import java.util.List;
 import java.util.stream.Stream;
-import lotto.lottoapp.value.LottoNumber;
-import lotto.lottoapp.value.LottoNumbers;
-import lotto.lottoapp.value.Won;
+import lotto.lottoapp.model.value.LottoNumber;
+import lotto.lottoapp.model.value.LottoNumbers;
+import lotto.lottoapp.model.value.Won;
 
 public class Lotto {
