+++ src/main/java/lotto/lottoapp/model/Lottos.java
@@ -8,10 +8,10 @@ import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
-import lotto.lottoapp.value.LottoNumbers;
-import lotto.lottoapp.value.WinningResult;
-import lotto.lottoapp.value.WinningStatistics;
-import lotto.lottoapp.value.Won;
+import lotto.lottoapp.model.value.LottoNumbers;
+import lotto.lottoapp.model.value.WinningResult;
+import lotto.lottoapp.model.value.WinningStatistics;
+import lotto.lottoapp.model.value.Won;
 
 public class Lottos {
