+++ src/main/java/lotto/lottoapp/model/value/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.lottoapp.value;
+package lotto.lottoapp.model.value;
 
 import java.util.Objects;
