+++ src/main/java/lotto/lottoapp/model/value/Won.java
@@ -1,4 +1,4 @@
-package lotto.lottoapp.value;
+package lotto.lottoapp.model.value;
 
 import java.math.BigInteger;
 import java.text.DecimalFormat;
