+++ src/test/java/lotto/model/LottosTest.java
@@ -1,10 +1,10 @@
 package lotto.model;
 
-import static lotto.lottoapp.value.WinningResult.FIFTH;
-import static lotto.lottoapp.value.WinningResult.FIRST;
-import static lotto.lottoapp.value.WinningResult.FOURTH;
-import static lotto.lottoapp.value.WinningResult.SECOND;
-import static lotto.lottoapp.value.WinningResult.THIRD;
+import static lotto.lottoapp.model.value.WinningResult.FIFTH;
+import static lotto.lottoapp.model.value.WinningResult.FIRST;
+import static lotto.lottoapp.model.value.WinningResult.FOURTH;
+import static lotto.lottoapp.model.value.WinningResult.SECOND;
+import static lotto.lottoapp.model.value.WinningResult.THIRD;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
@@ -13,10 +13,10 @@ import java.util.List;
 import lotto.lottoapp.model.AutomaticLottoNumbersGenerator;
 import lotto.lottoapp.model.Lottos;
 import lotto.lottoapp.model.WinningLotto;
-import lotto.lottoapp.value.BonusNumber;
-import lotto.lottoapp.value.LottoNumbers;
-import lotto.lottoapp.value.WinningStatistics;
-import lotto.lottoapp.value.Won;
+import lotto.lottoapp.model.value.BonusNumber;
+import lotto.lottoapp.model.value.LottoNumbers;
+import lotto.lottoapp.model.value.WinningStatistics;
+import lotto.lottoapp.model.value.Won;
 import org.junit.jupiter.api.Test;
 
 class LottosTest {
