+++ src/test/java/lotto/model/WinningLottoTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import lotto.lottoapp.model.WinningLotto;
-import lotto.lottoapp.value.BonusNumber;
-import lotto.lottoapp.value.LottoNumbers;
+import lotto.lottoapp.model.value.BonusNumber;
+import lotto.lottoapp.model.value.LottoNumbers;
 import org.junit.jupiter.api.Test;
 
 class WinningLottoTest {
